Get rid of Config()
The difference between a config dir and a data dir is clear, but not portable.
Nowadays, apps like Chromium and Syncthing store all of their data and config
in a single directory.

On linux, this ends up being ~/.config. The real data directory as specified
by XDG, ~/.local/share, is rarely used by any apps.

Since other OSes like windows have an APPDATA dir but not an APPCONFIG one,
it's best to merge the config and data dirs.